alshamelah_api/requirements.txt
Django==3.0.7
pyyaml==5.3.1
PyArabic==0.6.7
# for pdf generation with arabic support
# --------------------------------------
# reportlab==3.5.42
# arabic_reshaper==2.0.15
# python-bidi==0.4.2
# --------------------------------------
django-allauth==0.42.0
django-crispy-forms==1.9.1
django-debug-toolbar==2.2
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
django-rest-auth==0.9.5
django-role-permissions==3.0.0
django-coreapi==2.3.0
django-filter==2.2.0
django-model-utils==4.0.0
django-extensions==2.2.9
#django-rest-swagger==2.2.0
drf-yasg==1.17.1
drf-nested-routers==0.91.0
drf-extensions==0.6
Pillow==7.0.0
voluptuous==0.11.7  # JSON validation
whitenoise==5.0.1
postgres==3.0.0

gunicorn==19.9.0
psycopg2-binary==2.8.3
munch==2.5.0
coreapi==2.3.3
easy-thumbnails==2.7
django-cleanup==5.0.0
twilio==6.43
python-dateutil==2.8.1